# backend/node/docs/openapi.yaml
openapi: 3.1.0
info:
  title: GymPoint API
  version: 1.0.0
  description: API del sistema GymPoint para gestión de gimnasios, rutinas, entrenamientos y gamificación.
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Datos inválidos o request malformada
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Datos inválidos
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Recurso no encontrado
...
  parameters:
...
    IdPathParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: Identificador único del recurso
...
    ExerciseIdPathParam:
      name: id_exercise
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: ID del ejercicio
...
  schemas:
...
    Error:
      type: object
      required:
        - error
      additionalProperties: false
      properties:
        error:
          type: string
          description: Mensaje de error descriptivo
          example: Ha ocurrido un error
...
    GymSpecialScheduleResponse:
      type: object
      required:
        - id_special_schedule
        - id_gym
        - date
        - closed
        - motive
        - created_at
        - updated_at
      properties:
        id_special_schedule:
          type: integer
          example: 1
        id_gym:
          type: integer
          example: 5
        date:
          type: string
          format: date
          description: Fecha del horario especial
          example: '2025-12-25'
        opening_time:
          type:
            - string
            - 'null'
          format: time
          description: Hora de apertura (HH:MM)
          example: '10:00'
        closing_time:
          type:
            - string
            - 'null'
          format: time
          description: Hora de cierre (HH:MM)
          example: '14:00'
        closed:
          type: boolean
          description: Si el gimnasio está cerrado todo el día
          default: false
        motive:
          type: string
          description: Motivo del horario especial
          example: Feriado Nacional - Navidad
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateGymSpecialScheduleRequest:
      type: object
      additionalProperties: false
      required:
        - id_gym
        - date
        - closed
        - motive
      properties:
        id_gym:
          type: integer
          minimum: 1
        date:
          type: string
          format: date
        opening_time:
          type: string
          format: time
          pattern: ^([0-1][0-9]|2[0-3]):[0-5][0-9]$
        closing_time:
          type: string
          format: time
          pattern: ^([0-1][0-9]|2[0-3]):[0-5][0-9]$
        closed:
          type: boolean
        motive:
          type: string
          minLength: 1
          maxLength: 500
    UpdateGymSpecialScheduleRequest:
      type: object
      additionalProperties: false
      properties:
        id_gym:
          type: integer
          minimum: 1
        date:
          type: string
          format: date
        opening_time:
          type: string
          format: time
          pattern: ^([0-1][0-9]|2[0-3]):[0-5][0-9]$
        closing_time:
          type: string
          format: time
          pattern: ^([0-1][0-9]|2[0-3]):[0-5][0-9]$
        closed:
          type: boolean
        motive:
          type: string
          minLength: 1
          maxLength: 500
...
    ProgressResponse:
      type: object
      properties:
        id_progress:
          type: integer
          description: ID del registro de progreso
          example: 1
        id_user_profile:
          type: integer
          description: ID del perfil de usuario
          example: 10
        date:
          type: string
          format: date
          description: Fecha del registro
          example: '2025-01-15'
        total_weight_lifted:
          type: number
          format: float
          description: Peso total levantado en kg
          example: 5000
        total_reps:
          type: integer
          description: Repeticiones totales
          example: 100
        total_sets:
          type: integer
          description: Sets totales
          example: 20
        notes:
          type: string
          description: Notas adicionales
          example: Great workout session
        created_at:
          type: string
          format: date-time
          description: Fecha de creación
        updated_at:
          type: string
          format: date-time
          description: Fecha de última actualización
...
paths:
...
  /api/gym-special-schedules/{gymId}:
    get:
      operationId: listGymSpecialSchedules
      tags:
        - Gym Schedules
      summary: Listar horarios especiales de un gimnasio
      description: Obtiene todos los horarios especiales de un gimnasio específico
      security:
        - BearerAuth: []
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: integer
          description: ID del gimnasio
      responses:
        '200':
          description: Lista de horarios especiales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GymSpecialScheduleResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: createGymSpecialSchedule
      tags:
        - Gym Schedules
      summary: Crear un horario especial
      description: Crea un nuevo horario especial para el gimnasio (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: integer
          description: ID del gimnasio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGymSpecialScheduleRequest'
      responses:
        '201':
          description: Horario especial creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GymSpecialScheduleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/gym-special-schedules/{id}:
    put:
      operationId: updateGymSpecialSchedule
      tags:
        - Gym Schedules
      summary: Actualizar un horario especial
      description: Actualiza los datos de un horario especial (admin)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
          description: ID del horario especial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGymSpecialScheduleRequest'
      responses:
        '200':
          description: Horario especial actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GymSpecialScheduleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteGymSpecialSchedule
      tags:
        - Gym Schedules
      summary: Eliminar un horario especial
      description: Elimina un horario especial
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
          description: ID del horario especial
      responses:
        '204':
          description: Horario especial eliminado exitosamente
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/progress:
    post:
      operationId: registerProgress
      tags:
        - Progress
      summary: Registrar progreso físico del usuario
      description: |
        Registra el progreso físico del usuario incluyendo peso total levantado, repeticiones totales, sets totales y ejercicios realizados.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - date
                - total_weight_lifted
                - total_reps
                - total_sets
              properties:
                date:
                  type: string
                  format: date
                  description: Fecha del registro
                  example: '2025-01-15'
                total_weight_lifted:
                  type: number
                  format: float
                  description: Peso total levantado en kg
                  example: 5000
                total_reps:
                  type: integer
                  description: Repeticiones totales
                  example: 100
                total_sets:
                  type: integer
                  description: Sets totales
                  example: 20
                notes:
                  type: string
                  description: Notas adicionales
                  example: Great workout session
                exercises:
                  type: array
                  description: Lista de ejercicios realizados
                  items:
                    type: object
                    required:
                      - id_exercise
                      - used_weight
                      - reps
                    properties:
                      id_exercise:
                        type: integer
                        description: ID del ejercicio
                        example: 5
                      used_weight:
                        type: number
                        format: float
                        description: Peso utilizado en kg
                        example: 80
                      reps:
                        type: integer
                        description: Número de repeticiones
                        example: 10
                      sets:
                        type: integer
                        description: Número de sets
                        example: 3
                        default: 1
      responses:
        '201':
          description: Progreso registrado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Progreso registrado con éxito
                  data:
                    $ref: '#/components/schemas/ProgressResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/progress/me:
    get:
      operationId: getUserProgress
      tags:
        - Progress
      summary: Obtener historial de progreso del usuario
      description: |
        Obtiene todo el historial de progreso físico del usuario autenticado con opciones de paginación y filtrado.
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: Número máximo de registros a retornar
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Número de registros a saltar
        - name: include_exercises
          in: query
          schema:
            type: boolean
            default: false
          description: Incluir detalles de ejercicios
      responses:
        '200':
          description: Historial de progreso obtenido
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProgressResponse'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/progress/me/estadistica:
    get:
      operationId: getWeightStats
      tags:
        - Progress
      summary: Obtener estadísticas de peso levantado
      description: |
        Obtiene estadísticas de peso total levantado del usuario en un rango de fechas opcional.
      security:
        - BearerAuth: []
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
          description: Fecha de inicio del rango
          example: '2025-01-01'
        - name: end_date
          in: query
          schema:
            type: string
            format: date
          description: Fecha de fin del rango
          example: '2025-01-31'
      responses:
        '200':
          description: Estadísticas obtenidas
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        date:
                          type: string
                          format: date
                        total_weight_lifted:
                          type: number
                        total_reps:
                          type: integer
                        total_sets:
                          type: integer
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/progress/me/ejercicios:
    get:
      operationId: getExerciseHistory
      tags:
        - Progress
      summary: Obtener historial completo de ejercicios
      description: |
        Obtiene el historial completo de todos los ejercicios realizados por el usuario autenticado.
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
          description: Número máximo de registros
      responses:
        '200':
          description: Historial de ejercicios obtenido
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        date:
                          type: string
                          format: date
                        exercise_name:
                          type: string
                        muscular_group:
                          type: string
                        used_weight:
                          type: number
                        reps:
                          type: integer
                        sets:
                          type: integer
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/progress/me/ejercicios/{id_exercise}/mejor:
    get:
      operationId: getPersonalRecord
      tags:
        - Progress
      summary: Obtener récord personal de un ejercicio
      description: |
        Obtiene el mejor levantamiento (peso máximo) del usuario para un ejercicio específico.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExerciseIdPathParam'
          description: ID del ejercicio
          example: 5
      responses:
        '200':
          description: Récord personal obtenido
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      date:
                        type: string
                        format: date
                      used_weight:
                        type: number
                      reps:
                        type: integer
                      sets:
                        type: integer
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/progress/me/ejercicios/{id_exercise}/promedio:
    get:
      operationId: getExerciseAverages
      tags:
        - Progress
      summary: Obtener promedios de un ejercicio
      description: |
        Obtiene los promedios de peso, repeticiones y sets para un ejercicio específico del usuario.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExerciseIdPathParam'
          description: ID del ejercicio
          example: 5
      responses:
        '200':
          description: Promedios obtenidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      average_weight:
                        type: number
                        description: Peso promedio
                      average_reps:
                        type: number
                        description: Repeticiones promedio
                      average_sets:
                        type: number
                        description: Sets promedio
                      total_records:
                        type: integer
                        description: Total de registros
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
security:
  - bearerAuth: []